common knn graph build params
