+++ cpp/src/neighbors/all_neighbors/all_neighbors_builder.cuh
@@ -116,13 +116,19 @@ struct all_neighbors_builder_ivfpq : public all_neighbors_builder<T, IdxT> {
     size_t min_cluster_size,
     size_t max_cluster_size,
     size_t k,
-    all_neighbors::graph_build_params::ivf_pq_params& params,
+    graph_build_params::ivf_pq_params& params,
     std::optional<raft::device_matrix_view<IdxT, IdxT, row_major>> indices = std::nullopt,
     std::optional<raft::device_matrix_view<T, IdxT, row_major>> distances  = std::nullopt)
     : all_neighbors_builder<T, IdxT>(
         res, n_clusters, min_cluster_size, max_cluster_size, k, indices, distances),
       all_ivf_pq_params{params}
   {
+    if (params.refinement_rate != 2.0) {
+      RAFT_LOG_WARN(
+        "Recommend 2.0 for the refinement rate when using ivf_pq as the build algo. Setting "
+        "refinement rate to 2.0");
+      all_ivf_pq_params.refinement_rate = 2.0;
+    }
   }
 
   void prepare_build_common(size_t num_cols)
@@ -287,7 +293,7 @@ struct all_neighbors_builder_ivfpq : public all_neighbors_builder<T, IdxT> {
                        dataset_h.data_handle(), dataset.extent(0), dataset.extent(1)));
   }
 
-  all_neighbors::graph_build_params::ivf_pq_params all_ivf_pq_params;
+  graph_build_params::ivf_pq_params all_ivf_pq_params;
   size_t candidate_k;
 
   std::optional<raft::device_matrix<T, IdxT>> data_d;
@@ -308,7 +314,7 @@ struct all_neighbors_builder_nn_descent : public all_neighbors_builder<T, IdxT> 
     size_t min_cluster_size,
     size_t max_cluster_size,
     size_t k,
-    all_neighbors::graph_build_params::nn_descent_params& params,
+    graph_build_params::nn_descent_params& params,
     std::optional<raft::device_matrix_view<IdxT, IdxT, row_major>> indices = std::nullopt,
     std::optional<raft::device_matrix_view<T, IdxT, row_major>> distances  = std::nullopt)
     : all_neighbors_builder<T, IdxT>(
